Refactor project structure and rename packages
This commit renames the packages in the project to follow the standard Java naming conventions. The package names are now all in lowercase and separated by dots. Additionally, the commit also refactors the project structure to separate the source code from the compiled code.

The following files were renamed:
- Bokoblin.java -> bokwadventureonline.Bokoblin.java
- Character.java -> botwadventureonline.Character.java
- Client.java -> botwadventureonline.Client.java
- ClientThread.java -> botwadventureonline.ClientThread.java
- ExperiencePotion.java -> botwadventureonline.ExperiencePotion.java
- Field.java -> botwadventureonline.Field.java
- GameMap.java -> botwadventureonline.GameMap.java
- Guardian.java -> botwadventureonline.Guardian.java
- HealthPotion.java -> botwadventureonline.HealthPotion.java
- Hestu.java -> botwadventureonline.Hestu.java
- Inventory.java -> botwadventureonline.Inventory.java
- Item.java -> botwadventureonline.Item.java
- Korok.java -> botwadventureonline.Korok.java
- Lynel.java -> botwadventureonline.Lynel.java
- Merchant.java -> botwadventureonline.Merchant.java
- Moblin.java -> botwadventureonline.Moblin.java

In addition, the following files were deleted:
- Main.java
- SaveManager.java

And the following file was added:
- Main.java

These changes were made to improve the organization and maintainability of the project.